move time and temperature conversion functions to data_utils.py
