+++ src/netcdf_utils.py
@@ -113,37 +113,6 @@ def write_dataframe_to_netcdf(df, file_path):
         if dataset is not None:
             dataset.close()
 
-def time_to_datetime(time, reference_date='2018-04-14'):
-    """
-    Convert a dataframe time column to a datetime object.
-
-    Parameters:
-    time (pandas.Series): The time column to be converted.
-
-    Returns:
-    pandas.Series: The time column converted to a datetime object.
-
-    Example:
-    >>> df = pd.DataFrame({'time': [0, 1, 2]})
-    """
-    time = pd.to_timedelta(time, unit='h')
-    return pd.to_datetime(time + pd.to_datetime(reference_date))
-
-def celsius_to_kelvin(temp):
-    """
-    Convert a temperature in Celsius to Kelvin.
-
-    Parameters:
-    temp (float): The temperature in Celsius.
-
-    Returns:
-    float: The temperature in Kelvin.
-
-    Example:
-    >>> celsius_to_kelvin(0)
-    273.15
-    """
-    return temp + 273.15
 
 if __name__ == '__main__':
     # Example usage
@@ -159,13 +128,10 @@ if __name__ == '__main__':
     file_name = 'data/observation.nc'
     city_coordinates = (8,26)
     df = read_netcdf_to_dataframe(file_name, city_coordinates)
-    df["time"] = time_to_datetime(df["time"])
-    df["temperature"] = df["temperature"].apply(celsius_to_kelvin)
     print(df.head())
 
     # Using the read_netcdf_to_dataframe function to the "forecast.nc" file
     # file_name = 'data/forecast.nc'
     # city_coordinates = (8,26)
     # df = read_netcdf_to_dataframe(file_name, city_coordinates)
-    # df["time"] = time_to_datetime(df["time"])
     # print(df.head())
